+++ source/Vec.hpp
@@ -16,13 +16,13 @@ class Vec2D
     {
     } // delegate constructor, default value 0
 
-    Vec2D(double x, double y)
+    Vec2D(float x, float y)
     {
         Data[0] = x;
         Data[1] = y;
     }
 
-    Vec2D(const double init)
+    Vec2D(const float init)
     {
         for (size_t i = 0; i < VN; i++)
         {
@@ -31,7 +31,7 @@ class Vec2D
         }
     }
 
-    Vec2D(const std::array<double, VN> &Copy) // duplicate vector
+    Vec2D(const std::array<float, VN> &Copy) // duplicate vector
     {
         for (size_t i = 0; i < VN; i++)
         {
@@ -40,9 +40,9 @@ class Vec2D
         }
     }
 
-    double SizeSqr() const
+    float SizeSqr() const
     {
-        double sum = 0;
+        float sum = 0;
         for (size_t i = 0; i < VN; i++)
         {
             sum += sqr(Data[i]);
@@ -50,7 +50,7 @@ class Vec2D
         return sum;
     }
 
-    double Size() const
+    float Size() const
     {
         return std::sqrt(SizeSqr());
     }
@@ -82,7 +82,7 @@ class Vec2D
         return Ret;
     }
 
-    Vec2D operator/(const double Denom) const
+    Vec2D operator/(const float Denom) const
     {
         Vec2D Ret(Data);
         for (size_t i = 0; i < VN; i++)
@@ -92,7 +92,7 @@ class Vec2D
         return Ret;
     }
 
-    Vec2D operator*(const double Scale) const
+    Vec2D operator*(const float Scale) const
     {
         Vec2D Ret(Data);
         for (size_t i = 0; i < VN; i++)
@@ -102,10 +102,10 @@ class Vec2D
         return Ret;
     }
 
-    Vec2D Rotate(const double Angle) const
+    Vec2D Rotate(const float Angle) const
     {
-        const double rX = cos(Angle) * Data[0] - sin(Angle) * Data[1];
-        const double rY = sin(Angle) * Data[0] + cos(Angle) * Data[1];
+        const float rX = cos(Angle) * Data[0] - sin(Angle) * Data[1];
+        const float rY = sin(Angle) * Data[0] + cos(Angle) * Data[1];
         return Vec2D(rX, rY);
     }
     //////////// :ASSIGNMENT: //////////////
@@ -125,7 +125,7 @@ class Vec2D
         }
     }
 
-    void operator/=(const double Denom)
+    void operator/=(const float Denom)
     {
         for (size_t i = 0; i < VN; i++)
         {
@@ -133,7 +133,7 @@ class Vec2D
         }
     }
 
-    void operator*=(const double Scale)
+    void operator*=(const float Scale)
     {
         for (size_t i = 0; i < VN; i++)
         {
@@ -150,21 +150,21 @@ class Vec2D
         }
     }
 
-    void RotateInPlace(const double Angle)
+    void RotateInPlace(const float Angle)
     {
-        double Tmp1 = cos(Angle) * Data[0] - sin(Angle) * Data[1];
-        double Tmp2 = sin(Angle) * Data[0] + cos(Angle) * Data[1];
+        float Tmp1 = cos(Angle) * Data[0] - sin(Angle) * Data[1];
+        float Tmp2 = sin(Angle) * Data[0] + cos(Angle) * Data[1];
         Data[0] = Tmp1;
         Data[1] = Tmp2;
     }
 
     //////////// :GETTERS: //////////////
-    double operator[](const size_t i) const
+    float operator[](const size_t i) const
     {
         return Data[i];
     }
 
-    Vec2D LimitMagnitude(const double MaxVec)
+    Vec2D LimitMagnitude(const float MaxVec)
     {
         if (SizeSqr() > sqr(MaxVec))
         {
@@ -173,7 +173,7 @@ class Vec2D
         return (*this);
     }
     // Easiest if Data is public
-    std::array<double, VN> Data; /// TODO: template the size
+    std::array<float, VN> Data; /// TODO: template the size
 };
 
 inline std::ostream &operator<<(std::ostream &OutStream, const Vec2D &V)
