+++ source/cudaSimulator.cu
@@ -5,22 +5,23 @@
 #include <chrono>     // timing threads
 #include <omp.h>      // OpenMP
 #include <string>     // cout
-#include <vector>     // std::vector
 #include <unordered_map>
+#include <vector> // std::vector
 
 #include <cuda.h>
-#include <math.h>
 #include <cuda_runtime.h>
 #include <driver_functions.h>
+#include <math.h>
 
-struct GlobalBoidData {
+struct GlobalBoidData
+{
     size_t numBoids;
 
-    float* position;
-    float* velocity;
-    float* acceleration;
-    int* flockID; 
-    int* flockSize;
+    float *position;
+    float *velocity;
+    float *acceleration;
+    int *flockID;
+    int *flockSize;
 };
 
 __constant__ GlobalBoidData cuConstBoidData;
@@ -28,7 +29,7 @@ __constant__ ParamsStruct cuConstGlobalParams;
 
 /** @return the result of adding the inputs together */
 __device__ float2 operator+(const float2 &A, const float2 &B)
-{ 
+{
     return make_float2(A.x + B.x, A.y + B.y);
 }
 
@@ -39,40 +40,41 @@ __device__ float2 operator-(const float2 &A, const float2 &B)
 }
 
 /** @return the result of multiplying each element of A by v */
-__device__ float2 operator*(const float2 &A, const float v) 
+__device__ float2 operator*(const float2 &A, const float v)
 {
     return make_float2(A.x * v, A.y * v);
 }
 
 /** @return the result of dividing each element of A by v */
-__device__ float2 operator/(const float2 &A, const float v) 
+__device__ float2 operator/(const float2 &A, const float v)
 {
     return make_float2(A.x / v, A.y / v);
 }
 
 /** @return the square of the l2 norm of the specified vector */
-__device__ float sizeSqrd (const float2 &A) 
+__device__ float sizeSqrd(const float2 &A)
 {
-    return A.x*A.x + A.y*A.y;
+    return A.x * A.x + A.y * A.y;
 }
 
 /** @return whether or not A lies beyond a distance of radius R from B */
-__device__ bool distGT (const float2 &A, const float2 &B, const float R)
+__device__ bool distGT(const float2 &A, const float2 &B, const float R)
 {
-    return sizeSqrd(A-B) > (R*R);
+    return sizeSqrd(A - B) > (R * R);
 }
 
 /** @return whether or not A lies within a distance of radius R from B */
-__device__ bool distLT (const float2 &A, const float2 &B, const float R)
+__device__ bool distLT(const float2 &A, const float2 &B, const float R)
 {
-    return sizeSqrd(A-B) < R*R;
+    return sizeSqrd(A - B) < R * R;
 }
 
-__device__ float2 normalize(const float2 &A) {
-    return A / (sqrt(double(sizeSqrd(A))));
+__device__ float2 normalize(const float2 &A)
+{
+    return A / (sqrt(float(sizeSqrd(A))));
 }
 
-__device__ float2 limitMagnitudeKernel(const float2 &A, const double maxMag)
+__device__ float2 limitMagnitudeKernel(const float2 &A, const float maxMag)
 {
     if (sizeSqrd(A) > maxMag * maxMag)
     {
@@ -80,44 +82,45 @@ __device__ float2 limitMagnitudeKernel(const float2 &A, const double maxMag)
     }
     return A;
 }
-/** 
+/**
  * Computes the acceleration for the boid indexed by index in the current
  * tick.
- */ 
-__global__ void 
-senseAndPlanKernel() 
+ */
+__global__ void senseAndPlanKernel()
 {
     const int index = (blockIdx.x * blockDim.x) + threadIdx.x;
 
-    if (index >= cuConstBoidData.numBoids) return;
+    if (index >= cuConstBoidData.numBoids)
+        return;
 
     cuConstBoidData.flockID[index] = 1;
     int N = cuConstBoidData.numBoids;
-    float2 a1 = make_float2(0.0,0.0); 
-    float2 a2 = make_float2(0.0,0.0); 
-    float2 a3 = make_float2(0.0,0.0);
-    float2* position = (float2*)cuConstBoidData.position;
-    float2* velocity = (float2*)cuConstBoidData.velocity;
-    float2* acceleration = (float2*)cuConstBoidData.acceleration;
+    float2 a1 = make_float2(0.0, 0.0);
+    float2 a2 = make_float2(0.0, 0.0);
+    float2 a3 = make_float2(0.0, 0.0);
+    float2 *position = (float2 *)cuConstBoidData.position;
+    float2 *velocity = (float2 *)cuConstBoidData.velocity;
+    float2 *acceleration = (float2 *)cuConstBoidData.acceleration;
 
     BoidParamsStruct boidParams = cuConstGlobalParams.BoidParams;
 
     float2 posUs = position[index];
     float2 velUs = velocity[index];
 
-    float2 relCOM = make_float2(0.0,0.0);  
-    float2 relCOV = make_float2(0.0,0.0);
-    float2 sep = make_float2(0.0,0.0);
+    float2 relCOM = make_float2(0.0, 0.0);
+    float2 relCOV = make_float2(0.0, 0.0);
+    float2 sep = make_float2(0.0, 0.0);
 
     int numCloseBy = 0;
-    for(int i = 0; i < N; i++) {
+    for (int i = 0; i < N; i++)
+    {
         float2 posThem = position[i];
         float2 velThem = velocity[i];
-        if (i != index && distLT(posUs,posThem,boidParams.NeighbourhoodRadius))
+        if (i != index && distLT(posUs, posThem, boidParams.NeighbourhoodRadius))
         {
-            relCOM = relCOM + posThem;  
+            relCOM = relCOM + posThem;
             relCOV = relCOV + velThem;
-            if (distLT(posUs,posThem,boidParams.CollisionRadius))
+            if (distLT(posUs, posThem, boidParams.CollisionRadius))
             {
                 sep = sep - (posThem - posUs);
             }
@@ -131,29 +134,28 @@ senseAndPlanKernel()
         a2 = sep * (boidParams.Separation); // dosent depent on NumCloseby but makes sense
         a3 = ((relCOV / numCloseBy) - velUs) * (boidParams.Alignment);
     }
-    
+
     acceleration[index] = a1 + a2 + a3;
     cuConstBoidData.flockID[index] = 1;
 }
 
-__global__ void 
-actKernel(double deltaTime)
+__global__ void actKernel(float deltaTime)
 {
     const int index = (blockIdx.x * blockDim.x) + threadIdx.x;
 
-    if (index >= cuConstBoidData.numBoids) return;
+    if (index >= cuConstBoidData.numBoids)
+        return;
 
     /// NOTE: This function is meant to be independent from all other boids
     /// and thus can be run asynchronously, however it needs a barrier between itself
     /// and the senseAndPlan() device function
-    float2* position = (float2*)cuConstBoidData.position;
-    float2* velocity = (float2*)cuConstBoidData.velocity;
-    const float2* acceleration = (float2*)cuConstBoidData.acceleration;
+    float2 *position = (float2 *)cuConstBoidData.position;
+    float2 *velocity = (float2 *)cuConstBoidData.velocity;
+    const float2 *acceleration = (float2 *)cuConstBoidData.acceleration;
 
     BoidParamsStruct params = cuConstGlobalParams.BoidParams;
 
-    velocity[index] = limitMagnitudeKernel((velocity[index] + acceleration[index]), 
-                                            params.MaxVel);
+    velocity[index] = limitMagnitudeKernel((velocity[index] + acceleration[index]), params.MaxVel);
     position[index] = position[index] + (velocity[index] * deltaTime);
 }
 
@@ -161,18 +163,18 @@ class Simulator
 {
   public:
     GlobalBoidData cuDevicePtrs;
-    float* cudaDevicePosition;
-    float* cudaDeviceVelocity;
-    float* cudaDeviceAcceleration;
-    int* cudaDeviceFlockID;
-    int* cudaDeviceFlockSize;
+    float *cudaDevicePosition;
+    float *cudaDeviceVelocity;
+    float *cudaDeviceAcceleration;
+    int *cudaDeviceFlockID;
+    int *cudaDeviceFlockSize;
 
     int numBoids;
-    float* position;
-    float* velocity;
-    float* acceleration;
-    int* flockID;
-    int* flockSize;
+    float *position;
+    float *velocity;
+    float *acceleration;
+    int *flockID;
+    int *flockSize;
     Simulator()
     {
         Params = GlobalParams.SimulatorParams;
@@ -208,7 +210,7 @@ class Simulator
 
     void Simulate()
     {
-        double ElapsedTime = 0;
+        float ElapsedTime = 0;
         for (size_t i = 0; i < Params.NumIterations; i++)
         {
             ElapsedTime += Tick();
@@ -221,26 +223,28 @@ class Simulator
      * Inserts the value v in the float2 array represented by arr
      * @pre i < |arr|/2
      */
-    void insertFloat2(float* arr, const Vec2D &v, const size_t i) {
-        arr[2*i] = v[0];
-        arr[2*i+1] = v[1];
+    void insertFloat2(float *arr, const Vec2D &v, const size_t i)
+    {
+        arr[2 * i] = v[0];
+        arr[2 * i + 1] = v[1];
     }
 
-    void InitBoidDataArrays() {
+    void InitBoidDataArrays()
+    {
         std::vector<Boid> &AllBoids = *(AllFlocks.begin()->second.Neighbourhood.GetAllBoidsPtr());
         size_t vecSize = sizeof(float) * 2 * numBoids;
         size_t intSize = sizeof(int) * numBoids;
 
-        position = (float*)malloc(vecSize);
-        velocity = (float*)malloc(vecSize);
-        acceleration = (float*)malloc(vecSize);
-        flockID = (int*)malloc(intSize);
-        flockSize = (int*)malloc(intSize);
+        position = (float *)malloc(vecSize);
+        velocity = (float *)malloc(vecSize);
+        acceleration = (float *)malloc(vecSize);
+        flockID = (int *)malloc(intSize);
+        flockSize = (int *)malloc(intSize);
 
-        for (size_t i=0; i < AllBoids.size(); i++)
+        for (size_t i = 0; i < AllBoids.size(); i++)
         {
             const Boid &boid = AllBoids[i];
-            
+
             insertFloat2(position, boid.Position, i);
             insertFloat2(velocity, boid.Velocity, i);
             insertFloat2(acceleration, boid.Acceleration, i);
@@ -252,22 +256,24 @@ class Simulator
     /**
      * @pre All Boid arrays have been initialized and are up to date.
      */
-    void UpdateBoidPosAndVel() {
+    void UpdateBoidPosAndVel()
+    {
         std::vector<Boid> &AllBoids = *(AllFlocks.begin()->second.Neighbourhood.GetAllBoidsPtr());
         for (size_t i = 0; i < AllBoids.size(); i++)
         {
             Boid &B = AllBoids[i];
-            B.Position = Vec2D(position[2*i],position[2*i+1]);
-            B.Velocity = Vec2D(velocity[2*i],velocity[2*i+1]);
+            B.Position = Vec2D(position[2 * i], position[2 * i + 1]);
+            B.Velocity = Vec2D(velocity[2 * i], velocity[2 * i + 1]);
         }
     }
 
-    /** 
+    /**
      * Sets up the memory required for the cuda device code
      *
-     * @return The time it took to setup. 
+     * @return The time it took to setup.
      */
-    double Setup() {
+    float Setup()
+    {
         auto StartTime = std::chrono::system_clock::now();
 
         // allocate boid and flock memory
@@ -282,16 +288,16 @@ class Simulator
         cudaMalloc(&flockSize, intSize);
 
         InitBoidDataArrays();
-        
-        cudaMemcpy(cudaDevicePosition,position,vecSize,cudaMemcpyHostToDevice);
-        cudaMemcpy(cudaDeviceVelocity,velocity,vecSize,cudaMemcpyHostToDevice);
-        cudaMemcpy(cudaDeviceAcceleration,acceleration,vecSize,cudaMemcpyHostToDevice);
-        cudaMemcpy(cudaDeviceFlockID,flockID,intSize,cudaMemcpyHostToDevice);
-        cudaMemcpy(cudaDeviceFlockSize,flockSize,intSize,cudaMemcpyHostToDevice);
+
+        cudaMemcpy(cudaDevicePosition, position, vecSize, cudaMemcpyHostToDevice);
+        cudaMemcpy(cudaDeviceVelocity, velocity, vecSize, cudaMemcpyHostToDevice);
+        cudaMemcpy(cudaDeviceAcceleration, acceleration, vecSize, cudaMemcpyHostToDevice);
+        cudaMemcpy(cudaDeviceFlockID, flockID, intSize, cudaMemcpyHostToDevice);
+        cudaMemcpy(cudaDeviceFlockSize, flockSize, intSize, cudaMemcpyHostToDevice);
 
         // set up constants struct for copying into device memory struct
         cuDevicePtrs.numBoids = numBoids;
-        cuDevicePtrs.position = cudaDevicePosition; 
+        cuDevicePtrs.position = cudaDevicePosition;
         cuDevicePtrs.velocity = cudaDeviceVelocity;
         cuDevicePtrs.acceleration = cudaDeviceAcceleration;
         cuDevicePtrs.flockID = cudaDeviceFlockID;
@@ -301,12 +307,12 @@ class Simulator
         cudaMemcpyToSymbol(cuConstGlobalParams, &GlobalParams, sizeof(ParamsStruct));
 
         auto EndTime = std::chrono::system_clock::now();
-        std::chrono::duration<double> ElapsedTime = EndTime - StartTime;
+        std::chrono::duration<float> ElapsedTime = EndTime - StartTime;
 
         return ElapsedTime.count(); // return wall clock time diff
     }
 
-    double Tick()
+    float Tick()
     {
         // Run our actual problem (boid computation)
         auto StartTime = std::chrono::system_clock::now();
@@ -314,28 +320,22 @@ class Simulator
         const int threadsPerBlock = 64;
         const int numBlocks = (numBoids + threadsPerBlock - 1) / threadsPerBlock;
 
-        senseAndPlanKernel<<<numBlocks,threadsPerBlock>>>();
+        senseAndPlanKernel<<<numBlocks, threadsPerBlock>>>();
         cudaDeviceSynchronize();
-        actKernel<<<numBlocks,threadsPerBlock>>>(Params.DeltaTime);
+        actKernel<<<numBlocks, threadsPerBlock>>>(Params.DeltaTime);
         cudaDeviceSynchronize();
 
-        cudaMemcpy(position, 
-                   cuDevicePtrs.position,
-                   numBoids * sizeof(float) * 2,
-                   cudaMemcpyDeviceToHost);
-        cudaMemcpy(velocity, 
-                   cuDevicePtrs.velocity,
-                   numBoids * sizeof(float) * 2,
-                   cudaMemcpyDeviceToHost);
+        cudaMemcpy(position, cuDevicePtrs.position, numBoids * sizeof(float) * 2, cudaMemcpyDeviceToHost);
+        cudaMemcpy(velocity, cuDevicePtrs.velocity, numBoids * sizeof(float) * 2, cudaMemcpyDeviceToHost);
 
-        UpdateBoidPosAndVel();     
+        UpdateBoidPosAndVel();
         std::vector<Flock *> AllFlockPtrs = GetAllFlocksVector();
-        
+
         UpdateFlocks(AllFlockPtrs);
 
         auto EndTime = std::chrono::system_clock::now();
 
-        std::chrono::duration<double> ElapsedTime = EndTime - StartTime;
+        std::chrono::duration<float> ElapsedTime = EndTime - StartTime;
         // save tracer data
         Tracer::AddTickT(ElapsedTime.count());
 
@@ -347,7 +347,7 @@ class Simulator
 
         return ElapsedTime.count(); // return wall clock time diff
     }
-    
+
     std::vector<Flock *> GetAllFlocksVector() const
     {
         std::vector<Flock *> AllFlocksVec;
@@ -365,8 +365,9 @@ class Simulator
     {
 #pragma omp parallel num_threads(Params.NumThreads) // spawns threads
         {
-            if (GlobalParams.FlockParams.UseFlocks) {
-            /// NOTE: the following parallel operations are per-flocks, not per-boids
+            if (GlobalParams.FlockParams.UseFlocks)
+            {
+                /// NOTE: the following parallel operations are per-flocks, not per-boids
 #pragma omp for schedule(static)
                 for (size_t i = 0; i < AllFlocksVec.size(); i++)
                 {
@@ -410,7 +411,6 @@ class Simulator
 SimulatorParamsStruct Simulator::Params;
 ImageParamsStruct Image::Params;
 TracerParamsStruct Tracer::Params;
-
 
 // global params struct
 ParamsStruct GlobalParams;
